test: add smoke tests for converted modules
